Rename mobilefilelist file
