+++ frontend/components/FileTable/index.tsx
@@ -1,6 +1,6 @@
 import { File } from "../../lib/generated/graphql";
 import { FileTable } from "./FileTable";
-import { MobileFileList } from "./MobileFileTable";
+import { MobileFileList } from "./MobileFileList";
 
 export { MobileFileList, FileTable };
 export type { File };
